ii
git-svn-id: svn://192.168.0.99/MyUtility@7290 5ada3c48-002d-0410-850d-99865cc6fd0f

+++ CreateWString.h
@@ -30,9 +30,9 @@ inline static void ConvertToW(LPCSTR pIN, size_t inlen, LPWSTR* ppOut, size_t ou
 		CP_ACP,
 		0,
 		pIN,
-		inlen,
+		(int)inlen,
 		*ppOut,
-		outlen
+		(int)outlen
 		);
 }
